+++ cached_path/_cached_path.py
@@ -8,7 +8,7 @@ import tarfile
 import shutil
 
 from cached_path.cache_file import CacheFile
-from cached_path.common import PathOrStr, CACHE_DIRECTORY
+from cached_path.common import PathOrStr, get_cache_dir
 from cached_path.file_lock import FileLock
 from cached_path.meta import Meta
 from cached_path.protocols import get_cacher, hf_get_from_cache
@@ -24,7 +24,7 @@ logger = logging.getLogger("cached_path")
 
 def cached_path(
     url_or_filename: PathOrStr,
-    cache_dir: PathOrStr = CACHE_DIRECTORY,
+    cache_dir: Optional[PathOrStr] = None,
     extract_archive: bool = False,
     force_extract: bool = False,
 ) -> str:
@@ -34,45 +34,72 @@ def cached_path(
     then return the path to the cached file. If it's already a local path,
     make sure the file exists and return the path.
 
-    For URLs, "http://", "https://", "s3://", "gs://", and "hf://" are all supported.
-    The latter corresponds to the HuggingFace Hub.
+    For URLs, the following protocols are all supported:
 
-    For example, to download the PyTorch weights for the model `epwalsh/bert-xsmall-dummy`
-    on HuggingFace, you could do:
+    * ``http://`` and ``https://``,
+    * ``s3://`` for objects on `AWS S3`_,
+    * ``gs://`` for objects on `Google Cloud Storage (GCS)`_, and
+    * ``hf://`` for objects or repositories on `HuggingFace Hub`_.
 
-    ```python
-    cached_path("hf://epwalsh/bert-xsmall-dummy/pytorch_model.bin")
-    ```
+    .. _AWS S3: https://aws.amazon.com/s3/
+    .. _Google Cloud Storage (GCS): https://cloud.google.com/storage
+    .. _HuggingFace Hub: https://huggingface.co/
 
-    For paths or URLs that point to a tarfile or zipfile, you can also add a path
-    to a specific file to the `url_or_filename` preceeded by a "!", and the archive will
-    be automatically extracted (provided you set `extract_archive` to `True`),
-    returning the local path to the specific file. For example:
+    Examples
+    --------
 
-    ```python
-    cached_path("model.tar.gz!weights.th", extract_archive=True)
-    ```
+    To download a file over ``https``::
 
-    # Parameters
+        cached_path("https://github.com/allenai/cached_path/blob/main/README.md")
 
-    url_or_filename : `PathOrStr`
+    To download an object on GCS::
+
+        cached_path("gs://allennlp-public-models/lerc-2020-11-18.tar.gz")
+
+    To download the PyTorch weights for the model `epwalsh/bert-xsmall-dummy`_
+    on HuggingFace, you could do::
+
+        cached_path("hf://epwalsh/bert-xsmall-dummy/pytorch_model.bin")
+
+    For paths or URLs that point to a tarfile or zipfile, you can append the path
+    to a specific file within the archive to the ``url_or_filename``, preceeded by a "!".
+    The archive will be automatically extracted (provided you set ``extract_archive`` to ``True``),
+    returning the local path to the specific file. For example::
+
+        cached_path("model.tar.gz!weights.th", extract_archive=True)
+
+    .. _epwalsh/bert-xsmall-dummy: https://huggingface.co/epwalsh/bert-xsmall-dummy
+
+    Parameters
+    ----------
+
+    url_or_filename :
         A URL or path to parse and possibly download.
 
-    cache_dir : `PathOrStr`, optional (default = `CACHE_DIRECTORY`)
-        The directory to cache downloads.
+    cache_dir :
+        The directory to cache downloads. If not specified, the global default cache directory
+        will be used (``~/.cache/cached_path``). This can be set to something else with
+        :func:`set_cache_dir()`.
 
-    extract_archive : `bool`, optional (default = `False`)
-        If `True`, then zip or tar.gz archives will be automatically extracted.
+    extract_archive :
+        If ``True``, then zip or tar.gz archives will be automatically extracted.
         In which case the directory is returned.
 
-    force_extract : `bool`, optional (default = `False`)
-        If `True` and the file is an archive file, it will be extracted regardless
+    force_extract :
+        If ``True`` and the file is an archive file, it will be extracted regardless
         of whether or not the extracted directory already exists.
 
-        !!! Warning
+        .. important::
             Use this flag with caution! This can lead to race conditions if used
             from multiple processes on the same file.
+
+    Returns
+    -------
+    ``str``
+        The local path to the (potentially cached) resource.
+
     """
+    cache_dir = cache_dir if cache_dir else get_cache_dir()
     cache_dir = os.path.expanduser(cache_dir)
     os.makedirs(cache_dir, exist_ok=True)
 
@@ -198,11 +225,13 @@ def cached_path(
     return file_path
 
 
-def get_from_cache(url: str, cache_dir: PathOrStr = CACHE_DIRECTORY) -> str:
+def get_from_cache(url: str, cache_dir: Optional[PathOrStr] = None) -> str:
     """
     Given a URL, look for the corresponding dataset in the local cache.
     If it's not there, download it. Then return the path to the cached file.
     """
+    cache_dir = cache_dir if cache_dir else get_cache_dir()
+
     if url.startswith("hf://"):
         return hf_get_from_cache(url, cache_dir)