+++ cached_path/tqdm.py
@@ -9,6 +9,8 @@ import sys
 from time import time
 from typing import Optional
 
+from cached_path.common import FILE_FRIENDLY_LOGGING
+
 try:
     SHELL = str(type(get_ipython()))  # type:ignore # noqa: F821
 except:  # noqa: E722
@@ -80,7 +82,7 @@ class Tqdm:
     @staticmethod
     def tqdm(*args, **kwargs):
         # Use a slower interval when FILE_FRIENDLY_LOGGING is set.
-        default_mininterval = 2.0 if os.environ.get("FILE_FRIENDLY_LOGGING", False) else 0.1
+        default_mininterval = 2.0 if FILE_FRIENDLY_LOGGING else 0.1
 
         new_kwargs = {
             "file": TqdmToLogsWriter(),